requirements.txt
streamlit>=1.24.0
Pillow>=9.4.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
pdf2image>=1.16.3
python-docx>=0.8.11
reportlab>=3.6.12
openai>=1.3.0
python-dotenv>=1.0.0
python-telegram-bot>=20.8
twilio>=8.5.0
Flask>=2.3.0
requests>=2.32.2
